Areca: good disks assigned to a nonexistent state are 'Free'
They're not, really. What's probably happening is that there's not
enough disks that used to belong in the set for the controller to
construct the volume descriptor. In that case, this disk is not going to
be used for anything, but you've also got a controller or disk array in,
well, disarray.

Probably the right thing is to have an "AssignedTo" role for physical
disks, where we can indicate the volume its assigned to seperately from
its running state. That needs proper linkage between virtual and
physical disk objects though, which is on my todo list.