SASMPT: Ignore SIGPIPE during detection.
If there's a card, lsiutil will wait at the prompt. If not, it exits
immediately, but sometimes we're faster than it.

We could read the output channel before even trying to send anything, to
see if there's any controllers listed in the output. Or we could sleep.
Or any number of other ways.

The thing is, we don't really care: this entire thing succeeds and we
have controllers, or it doesn't and we don't. So its easier just to
remove roadblocks, go hard and see where we end up.

So here we are, removing a roadblock.